+++ app/Http/Controllers/api/v1/SettingsController.php
@@ -189,6 +189,7 @@ class SettingsController extends Controller
         $roomSettings->hide_owner_from_guests = $request->boolean('room_hide_owner_from_guests');
 
         $userSettings->password_change_allowed = $request->boolean('user_password_change_allowed');
+        $userSettings->search_by_name = $request->boolean('user_search_by_name');
 
         $bannerSettings->enabled = $request->boolean('banner_enabled');
         $bannerSettings->title = $request->input('banner_title');